* src/FileSystem.cs, src/Base/CameraFilesystem.cs: That should be   ulong, not 'long'
git-svn-id: https://svn.code.sf.net/p/gphoto/code/trunk/bindings/libgphoto2-sharp@10857 67ed7778-7388-44ab-90cf-0a291f65f57c

+++ src/FileSystem.cs
@@ -104,7 +104,7 @@ namespace Gphoto2
 		public long Capacity
 		{
 			get { return HasField(LibGPhoto2.CameraStorageInfoFields.MaxCapacity)
-				? storage.capacitykbytes * 1024 : -1; }
+				? (long)storage.capacitykbytes * 1024 : -1; }
 		}
 		
 		/// <value>
@@ -131,7 +131,7 @@ namespace Gphoto2
 		public long FreeSpace
 		{
 			get { return HasField(LibGPhoto2.CameraStorageInfoFields.FreeSpaceKbytes)
-				? storage.freekbytes * 1024 : -1; }
+				? (long)storage.freekbytes * 1024 : -1; }
 		}
 		
 		/// <value>
